back: feat!: add dtos to response, update entities ids generators, change structure of Action.java (remove json part and change it to enum fields), change ILobby and LobbyService by adding request dto to generateQuestion method. Add method in QuestionRepository for finding action

+++ Backend/src/main/java/com/telegram/app/backend/core/logic/dto/DefaultLobbyDto.java
@@ -14,10 +14,10 @@ public class DefaultLobbyDto {
     private SettingsMode gameType;
     @NotNull
     private List<DefaultPlayerDto> players;
-    private Integer difficultyFrom;
-    private Integer difficultyTo;
+    private Short difficultyFrom;
+    private Short difficultyTo;
     @NotNull
-    private List<Integer> packages;
+    private List<Short> packages;
 
     // Getters and Setters
 }
